Refactor: 클래스 및 DTO의 이름 명확성 증진을 위한 Refactoring
